requirements.txt
sentence-transformers>=2.2.2
numpy>=1.22.0,<2.0.0
pandas>=1.3.5
matplotlib>=3.5.1
plotly>=5.5.0
scikit-learn>=1.0.2
umap-learn>=0.5.2
nltk>=3.7
PyYAML>=6.0
PyPDF2>=2.0.0
